# 7171 CartoDBfy tables before switching in Sync Tables