--- a/app/models/synchronization/adapter.rb
+++ b/app/models/synchronization/adapter.rb
@@ -202,7 +202,7 @@ module CartoDB
       def import_cleanup(schema_name, table_name)
         qualified_table_name = "\"#{schema_name}\".#{table_name}"
 
-        user.in_database_direct_connection(statement_timeout: STATEMENT_TIMEOUT) do |user_database|
+        user.transaction_with_timeout(statement_timeout: STATEMENT_TIMEOUT) do |user_database|
           # When tables are created using ogr2ogr they are added a ogc_fid or gid primary key
           # In that case:
           #  - If cartodb_id already exists, remove ogc_fid