--- a/db/migrate/20160516103200_drop_visualization_exports_user_id_null_constraint.rb
+++ b/db/migrate/20160516103200_drop_visualization_exports_user_id_null_constraint.rb
@@ -1,0 +1,16 @@
+Sequel.migration do
+  up do
+    Rails::Sequel.connection.run(%{
+      ALTER TABLE visualization_exports alter column user_id drop not null;
+    })
+  end
+
+  down do
+    # This is the symmetric action, but it's not safe (there could be null `user_id`s)
+    # and it's not needed for `up` to be safe (`up` code is idempotent).
+    #
+    # Rails::Sequel.connection.run(%{
+    #   ALTER TABLE visualization_exports alter column user_id set not null;
+    # })
+  end
+end