Fixes #7171 

## The problem
Currently, the synchronization workflow is as follows:
  1. The new table is imported, it is never registered
  2. The old and the new tables are switched. The old table is dropped.
  3. CartoDBfication, import cleanups (cartodb_id) and geom checks (multi/simple geoms) are performed

And this provokes several problems:
  * When the switch is performed, the new table is not CartoDBfied (no cartodb_id, no the_geom) so any maps that contain this table as a layer will crash. See #7171 for a specific example.
  * If there's an error in the CartoDBfication or the rest of the checks, the synchronization will leave a broken table. The old table has been dropped so there's no possible rollback. The maps will be broken.
  * Also, if there's an error in the CartoDBfication, when the Ghost Tables are triggered they will unregister the table, breaking completely the synchronization. This would remove any layer in which the disconnected table was involved too.

When working in the issue, I noticed that the current flow is working due to the fact that when the switch between tables is performed, the new table gets automatically registered (it reuses the register of the old table) so the operations in the Table model that require the Table object can be performed on them.

As a side-effect, as the new table was not CartoDBfied it would not be captured when the ghost tables were triggered.

## The solution
We had to move the CartoDBfy process (and the rest of the checks) before the switch is performed to ensure that when it happens, the new table will be completely compatible with the maps and the Editor.

By "moving" these processes before the switch, we no longer have a registered table to work with. This is the reason why we duplicated code from Table (`cartodbfy`, `set_the_geom_column!` and `import_cleanup`) and adapted it to work with unregistered tables. Basically in this approach I am getting the column information via PostgreSQL queries, instead of checking the schema of the Table.

This code duplication is intended to be only part of a first iteration. Some parts of these methods can be directly removed with a correct configuration of Ogr2Ogr or the original methods can moved outside of Table.rb to create some kind of utility that can be used for both registered and unregistered tables.

To be able to CartoDBfy the table before switching, we had to move the table to the user schema. Some internal code of the CartoDBfication performs queries on CDB_UserQuotaSize and other functions that are expected to live in the schema in which the table lives, and `cdb_importer`, which is the default schema in which new tables are created does not contain these.

We also had to work with Ghost Tables. As the table is CartoDBfied and lives in the user schema, it would be visible for the Ghost Tables task. We're adding a commit to ignore tables in the User DB that are named as our `importer` tables.

Now the process is as follows:

  1. The new table is imported
  2. It is moved to the user schema
  3. It is CartoDBfied + import_cleanup + geom fixes(cartodb_id, the_geom)
  4. It is switched by the old table of the user

In this scenario, if something goes wrong while manipulating the table (incorrect CartoDB ID, or multiple geometry types issues, which were reported by Jaak) an exception will be raised and the sync will stop before any switch is made.

## Testing

Output of some tests: https://gist.github.com/iriberri/42a1a44a47f9ee5e428901dcd4a125d5


Other similar branches I created just for test purposes (see gist above):
* `7171-cdbfy_in_syncs_b`
* `cdbfy_master_with_sleep`

### Timing checks
I got some timings for the query edited in the `ghost_tables_manager` for my own production user with 2564 tables in the User DB.

* Original query, no name checks:
  * Time: 240.853 ms
* Adapted query 1: `t.tablename !~ 'importer_\\w{32}'`
  * Time: 364.519 ms
* Adapted query 2 (the one currently commited): `t.tablename !~ '^importer_'`
  * Time: 277.863 ms
* Adapted query 3: `t.tablename !~ '^importer_\\w{32}'`
  * Time: 278.990 ms

====

Pending: 
* [x] drop table in failure scenario to not leave stale tables
* [x] test thoroughly for cartodbfycation exceptions and special geom types (multipoint, simple linestrings) to check postgres queries flow
* [x] get some numbers on ghost tables query times