@iriberri 2016-05-27
#<Sequel::DatabaseError: PG::Error: ERROR:  CDB(_CDB_Has_Usable_Primary_ID: multiple primary keys for table "importer_928dc3b4242411e69a3d080027880ca6" are not allowed): ALTER TABLE importer_928dc3b4242411e69a3d080027880ca6 ADD CONSTRAINT cartodb_id_pk PRIMARY KEY (cartodb_id), ADD CONSTRAINT cartodb_id_integer CHECK (cartodb_id::integer >=0);
CONTEXT:  SQL statement "SELECT _CDB_Error(sql, Format('_CDB_Has_Usable_Primary_ID: %s', SQLERRM))"
PL/pgSQL function _cdb_has_usable_primary_id(regclass) line 60 at PERFORM
SQL statement "SELECT _CDB_Has_Usable_Primary_ID(reloid)"
PL/pgSQL function _cdb_rewrite_table(regclass,text) line 50 at SQL statement
SQL statement "SELECT _CDB_Rewrite_Table(reloid, destschema)"
PL/pgSQL function cdb_cartodbfytable(text,regclass) line 51 at PERFORM


```
                                                     Table "public.importer_928dc3b4242411e69a3d080027880ca6"
   Column    |       Type        |                                          Modifiers                                          | Storage  | Stats target | Description 
-------------+-------------------+---------------------------------------------------------------------------------------------+----------+--------------+-------------
 ogc_fid     | integer           | not null default nextval('importer_928dc3b4242411e69a3d080027880ca6_ogc_fid_seq'::regclass) | plain    |              | 
 cartodb_id  | integer           |                                                                                             | plain    |              | 
 the_geom    | geometry          |                                                                                             | main     |              | 
 name        | character varying |                                                                                             | extended |              | 
 description | character varying |                                                                                             | extended |              | 
Indexes:
    "importer_928dc3b4242411e69a3d080027880ca6_pkey" PRIMARY KEY, btree (ogc_fid)
Has OIDs: no
Options: autovacuum_enabled=true, toast.autovacuum_enabled=true


```