@iriberri 2016-05-27
https://github.com/CartoDB/cartodb/blob/b9245329c03dfe52db025a2e49d4a87ace7325d6/app/models/table.rb#L301-L360