@iriberri 2016-05-31
@rafatower thank you very much for the feedback! I just applied the `cartodbfy` changes here: https://github.com/CartoDB/cartodb/pull/7543/commits/16ab6ce28e6bc801048ae7946ed5b1eafc75dcfc

Won't squash the commits yet so that these changes are isolated.