@rafatower 2016-05-31
Adding @gfiorav to the party, especially regarding this: https://github.com/CartoDB/cartodb/pull/7543/files#diff-ed41caec536553c86246a629a466a888

See the comment above about performance.

As commented in person, we know a better approach is to use CDB_TableMetadata and mark the tables there upon CDB_CartodbfyTable (and it would be more performant, checking against catalog oids rather than doing string comparisons) but we evaluated it and this solution is far simpler and effective FTM.